Applied patch from http://issues.apache.org/activemq/browse/AMQ-1361
git-svn-id: https://svn.apache.org/repos/asf/activemq/trunk@565382 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-optional/src/main/java/org/apache/activemq/benchmark/BenchmarkSupport.java
@@ -176,7 +176,7 @@ public class BenchmarkSupport {
                 times++;
             }
             if (times > 0) {
-                average = total / times;
+                average = total / (double) times;
             }
 
             System.out.println(getClass().getName() + " Processed: " + processed + " messages this second. Average: " + average);
